requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
pillow
drf-yasg
django-filter
daphne
channels
selenium
channels_redis
django-notifications-hq
python-decouple
channels-redis
agora-token-builder
django-autoslug
django-bootstrap4
django-crispy-forms
daphne
django-redis
gunicorn
uvicorn
django-cors-headers
celery
redis
drf-writable-nested
